Plugin Modernization
This plugin was initially built with version 3 as the targeted API version. Adjustments were made to get it to work as a version 5 plugin but significant amounts of deprecated methods and workflows were being used. My game needs a functional billing library that's up-to-date so I built this plugin from the ground up. The Google Play Billing Library has changed quite a bit since the plugin was first created so here are some of the top changes.

There are now three primary methods for starting a purchase that depend on the purchase type.

1. If it's a non-consumable purchase, (like a level that you unlock or removing ads) use the `purchaseNonConsumable` method.
2. If it's consumable, (like buying virtual currency or a potion) then use `purchaseConsumable.` A top reason why these should be separated is because this version of the API now supports the ability to purchase consumable items with varying quantities.
3. The last one which needed the most refactoring are subscriptions. Use `purchaseSubscription` which requires not only a `productId` but also a `planId` which is the tier of subscription that the user is choosing.

Of course there are a lot more changes but these are the top differences. Acknowledging purchases and querying for product details is about the same workflow those just needed revisions on what methods need to be called.